Basic R language syntax is in this file
This file is containing basic R syntax. R is a high level programming language, but it is specially used in statistical field. There are some similarity and dissimilarity between R and Python. To understand the syntax of R, just follow the comments or "#" signs. R does not support multiline comment. This code can be run on R console, R studio or Shell terminal with the code-
"Rscript filename.R".